Rename builder to serialize
